moved some code into new components
