+++ app/src/features/surveys/view/SurveySpatialTelemetryDataTable.tsx
@@ -1,0 +1,45 @@
+import { SurveyContext } from 'contexts/surveyContext';
+import dayjs from 'dayjs';
+import { useContext, useMemo } from 'react';
+import NoSurveySectionData from '../components/NoSurveySectionData';
+import { ICritterDeployment } from '../telemetry/ManualTelemetryList';
+import SurveySpatialDataTable from './SurveySpatialDataTable';
+
+const SurveySpatialTelemetryDataTable = () => {
+  const surveyContext = useContext(SurveyContext);
+  const flattenedCritterDeployments: ICritterDeployment[] = useMemo(() => {
+    const data: ICritterDeployment[] = [];
+    // combine all critter and deployments into a flat list
+    surveyContext.deploymentDataLoader.data?.forEach((deployment) => {
+      const critter = surveyContext.critterDataLoader.data?.find(
+        (critter) => critter.critter_id === deployment.critter_id
+      );
+      if (critter) {
+        data.push({ critter, deployment });
+      }
+    });
+    return data;
+  }, [surveyContext.critterDataLoader.data, surveyContext.deploymentDataLoader.data]);
+
+  const mapData = () => {
+    return flattenedCritterDeployments.map((item) => {
+      return [
+        `${item.critter.animal_id}`,
+        `${item.deployment.device_id}`,
+        `${dayjs(item.deployment.attachment_start).format('YYYY-MM-DD')}`,
+        `${dayjs(item.deployment.attachment_end).format('YYYY-MM-DD')}`
+      ];
+    });
+  };
+  return (
+    <>
+      {flattenedCritterDeployments.length > 0 ? (
+        <SurveySpatialDataTable tableHeaders={['Alias', 'Device ID', 'Start', 'End']} tableRows={mapData()} />
+      ) : (
+        <NoSurveySectionData text="No data available" paperVariant="outlined" />
+      )}
+    </>
+  );
+};
+
+export default SurveySpatialTelemetryDataTable;
